+++ lib/roar/xml.rb
@@ -35,10 +35,10 @@ module Roar
         # FIXME: this doesn't belong into the generic XML representer.
         {
           :as => :link,
-          :collection   => true,
           :class        => Hypermedia::Hyperlink,
           :extend       => XML::HyperlinkRepresenter,
           :exec_context => :decorator,
+          collection: true
           } # TODO: merge with JSON.
       end
     end