+++ test/hypermedia_test.rb
@@ -2,35 +2,31 @@ require 'test_helper'
 
 class HypermediaTest < MiniTest::Spec
   describe "inheritance" do
-    before do
-      module BaseRepresenter
-        include Roar::JSON
-        include Roar::Hypermedia
+    module BaseRepresenter
+      include Roar::JSON
+      include Roar::Hypermedia
 
-        link(:base) { "http://base" }
-      end
+      link(:base) { "http://base" }
+    end
 
-      module Bar
-        include Roar::JSON
-        include Roar::Hypermedia
+    module Bar
+      include Roar::JSON
+      include Roar::Hypermedia
 
-        link(:bar) { "http://bar" }
-      end
+      link(:bar) { "http://bar" }
+    end
 
-      module Foo
-        include Roar::JSON
-        include Roar::Hypermedia
-        include BaseRepresenter
-        include Bar
+    module Foo
+      include Roar::JSON
+      include Roar::Hypermedia
+      include BaseRepresenter
+      include Bar
 
-        link(:foo) { "http://foo" }
-      end
+      link(:foo) { "http://foo" }
     end
 
     it "inherits parent links" do
-      foo = Object.new.extend(Foo)
-
-      assert_equal "{\"links\":[{\"rel\":\"base\",\"href\":\"http://base\"},{\"rel\":\"bar\",\"href\":\"http://bar\"},{\"rel\":\"foo\",\"href\":\"http://foo\"}]}", foo.to_json
+      Object.new.extend(Foo).to_json.must_equal "{\"links\":[{\"rel\":\"base\",\"href\":\"http://base\"},{\"rel\":\"bar\",\"href\":\"http://bar\"},{\"rel\":\"foo\",\"href\":\"http://foo\"}]}"
     end
 
     it "inherits links from all mixed-in representers" do
@@ -56,7 +52,7 @@ class HypermediaTest < MiniTest::Spec
     describe "#from_json" do
       it "parses" do
         subject.from_json "{\"links\":[{\"rel\":\"self\",\"href\":\"//self\"}]}"
-        subject.links.must_equal({"self" => link("rel" => "self", "href" => "//self")})
+        subject.links.must_equal("self" => link("rel" => "self", "href" => "//self"))
       end
     end
 
@@ -79,6 +75,7 @@ class HypermediaTest < MiniTest::Spec
         end
 
         it "renders rel" do
+          # raise subject.inspect
           subject.to_json.must_equal "{\"links\":[{\"rel\":\"ns:self\",\"href\":\"//self\"}]}"
         end
       end
@@ -135,56 +132,3 @@ class HypermediaTest < MiniTest::Spec
   end
 end
 
-class HyperlinkTest < MiniTest::Spec
-  describe "Hyperlink" do
-    subject { link(:rel => "self", "href" => "http://self", "data-whatever" => "Hey, @myabc") }
-
-    it "accepts string keys in constructor" do
-      assert_equal "Hey, @myabc", subject.send("data-whatever")
-    end
-
-    it "responds to #rel" do
-      assert_equal "self", subject.rel
-    end
-
-    it "responds to #href" do
-      assert_equal "http://self", subject.href
-    end
-
-    it "responds to #replace with string keys" do
-      subject.replace("rel" => "next")
-      assert_equal nil, subject.href
-      assert_equal "next", subject.rel
-    end
-
-    it "responds to #each and implements Enumerable" do
-      assert_equal ["rel:self", "href:http://self", "data-whatever:Hey, @myabc"], subject.collect { |k,v| "#{k}:#{v}" }
-    end
-  end
-
-  describe "Config inheritance" do
-      it "doesn't mess up with inheritable_array" do  # FIXME: remove this test when uber is out.
-        OpenStruct.new.extend( Module.new do
-                  include Roar::JSON
-                  include(Module.new do
-                                      include Roar::JSON
-                                      include Roar::Hypermedia
-
-                                      property :bla
-
-                                      link( :self) {"bo"}
-
-                                      #puts "hey ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~"
-                                      #puts representable_attrs.inheritable_array(:links).inspect
-                                    end)
-
-
-                  #puts representable_attrs.inheritable_array(:links).inspect
-
-                  property :blow
-                  include Roar::Hypermedia
-                  link(:bla) { "boo" }
-                end).to_hash.must_equal({"links"=>[{"rel"=>"self", "href"=>"bo"}, {"rel"=>"bla", "href"=>"boo"}]})
-    end
-  end
-end